The cop can often do a solid assessment when the maximum indent is 8
Only when there are no smaller indents, a higher indent is needed to unequivocally recognize soft tabstops. Start storing the minimum indent and extend the logic for solid assessment.
ENH: Better handle integer overflow when rating and normalizing: Limit to MAX_INT instead of carrying on with negative ratings, or just use Float values when Vim has support for it.