After overriding wrong indent settings, use single configured highlight settings or query use which one to use
The flow so far assumed that highlighting should be done and that there's only a single configuration for it. Now we need to distinguish between g:indentconsistencycop_highlighting and g:IndentConsistencyCop_AltHighlighting. To only query when there actually are inconsistencies, the query has to be inside s:HighlightInconsistentIndents(); we solve that by passing in a List of highlightingConfig then.